pythonProject/requirements.txt
joblib
numpy
scikit-learn
scipy
threadpoolctl
streamlit
pillow
pandas